patch: layout on mobile (#245)
* patch: layout on mobile

Closes #244

## Description
Currently, on mobile devices, the homepage has layout issues which cause the app to look odd and have part of the content cut-off.

This PR addresses that by some slight CSS tweaks. The main culprit was a display flex in the monetize div section which is necessary for larger devices but not for small for the layout. Outside of that, some slight padding items were tweaked to improve the look on mobile.

* change: minimum css size in media monetize section for consistency

* change: adding new line after media queries